django/requirements.txt
Django
djangorestframework
psycopg2
requests
django-cors-headers